backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.9.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==3.2.3
django-filter==21.1
django-import-export==3.3.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
et-xmlfile==1.1.0
flake8==6.1.0
gunicorn==20.1.0
idna==3.6
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pathspec==0.11.2
Pillow==9.0.0
platformdirs==4.0.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0
reportlab==4.0.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
tablib==3.5.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
webcolors==1.11.1
xlrd==2.0.1
xlwt==1.3.0
